Test/edition view

See merge request polytechnique-montr-al/log2990/20243/equipe-201/LOG2990-201!39

+++ client/src/app/services/edit-page-services/mouse-handler.service.ts
@@ -29,6 +29,11 @@ export class MouseHandlerService {
                     this.mapManagerService.removeItem(this.draggedItemInitRow, this.draggedItemInitCol);
                     this.draggedItemInitCol = null;
                     this.draggedItemInitRow = null;
+                    if (this.draggedItemInitRow !== null && this.draggedItemInitCol !== null) {
+                        this.mapManagerService.removeItem(this.draggedItemInitRow, this.draggedItemInitCol);
+                        this.draggedItemInitCol = null;
+                        this.draggedItemInitRow = null;
+                    }
                 }
             }
         }
@@ -86,6 +91,8 @@ export class MouseHandlerService {
         const item = this.mapManagerService.currentMap.mapArray[rowIndex][colIndex].item;
 
         if (item !== Item.NONE) {
+            this.draggedItemInitRow = rowIndex;
+            this.draggedItemInitCol = colIndex;
             event.dataTransfer?.setData('itemType', conversionConsts.itemToStringMap[item]);
             this.draggedItemInitRow = rowIndex;
             this.draggedItemInitCol = colIndex;
@@ -104,8 +111,11 @@ export class MouseHandlerService {
             if (
                 this.draggedItemInitRow !== null &&
                 this.draggedItemInitCol !== null &&
+                this.draggedItemInitRow !== null &&
+                this.draggedItemInitCol !== null &&
                 this.mapManagerService.currentMap.mapArray[rowIndex][colIndex].item === Item.NONE
             ) {
+                this.mapManagerService.removeItem(this.draggedItemInitRow, this.draggedItemInitCol);
                 this.mapManagerService.removeItem(this.draggedItemInitRow, this.draggedItemInitCol);
             }
             const item = conversionConsts.stringToItemMap[itemString];
@@ -116,6 +126,8 @@ export class MouseHandlerService {
                 this.mapManagerService.addItem(rowIndex, colIndex, item);
             }
         }
+        this.draggedItemInitRow = null;
+        this.draggedItemInitCol = null;
         this.draggedItemInitRow = null;
         this.draggedItemInitCol = null;
     }
